Add template_lint script and run on Travis
This script will make sure that:

(1) The templates all are parsed by Template() without error
(2) There are no duplicate tag loads
(3) All of the loads are used

I was surprised that after a bit of Googling I couldn't find a package
to do this.

It's not a great solution but if it's buggy we can easily remove it.